+++ tests/Unit/NestedTypePrivacyTests/ParentType.php
@@ -1,0 +1,27 @@
+<?php
+
+declare(strict_types = 1);
+namespace Rebing\GraphQL\Tests\Unit\NestedTypePrivacyTests;
+
+use GraphQL\Type\Definition\Type;
+use Rebing\GraphQL\Support\Facades\GraphQL;
+use Rebing\GraphQL\Support\Type as GraphQLType;
+
+class ParentType extends GraphQLType
+{
+    protected $attributes = [
+        'name' => 'Parent',
+    ];
+
+    public function fields(): array
+    {
+        return [
+            'name' => [
+                'type' => Type::nonNull(Type::string()),
+            ],
+            'child' => [
+                'type' => GraphQL::type('Child'),
+            ],
+        ];
+    }
+}