+++ tests/Unit/WithTypeTests/WrapperType.php
@@ -6,9 +6,10 @@ namespace Rebing\GraphQL\Tests\Unit\WithTypeTests;
 use GraphQL\Type\Definition\ObjectType;
 use GraphQL\Type\Definition\Type as GraphQLType;
 use Illuminate\Support\Collection;
+use Rebing\GraphQL\Support\Contracts\WrapType;
 use Rebing\GraphQL\Support\Facades\GraphQL;
 
-class WrapperType extends ObjectType
+class WrapperType extends ObjectType implements WrapType
 {
     /**
      * @param string $typeName The type name defined in graphql.php configuration file.